Modified experience section;
Added the ability to inject html (safe because this is a static page);
Modified the description of the fracture labs experience;
Added a bulleted list for clear separation;
Modified visibility of "Relocation & Work Authorization" when printing. (visible above communications in browser but visible below skills when printing.
Removed European Union from "Relocation & Work Authorization" because the target area is north America (Canada and United States).